% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{calc_WLA}
\alias{calc_WLA}
\title{Compute the wasteload allocation (WLA) used for effluent limit calculation.}
\usage{
calc_WLA(criteria, background, Qrsw, Qeff)
}
\arguments{
\item{criteria}{Limiting water quality criterion. Must be in same units as background argument.}

\item{background}{Background pollutant concentration. Must be in same units as criteria argument.}

\item{Qrsw}{Upstream limiting/design receiving water flowrate. Must be in same units as Qeff argument. Flow arguments may be entered in ratio form.}

\item{Qeff}{Effluent limiting/design flowrate. Must be in same units as Qrsw argument.  Flow arguments may be entered in ratio form.}
}
\value{
WLA as numeric value in same units as criteria and background.
}
\description{
Compute the wasteload allocation (WLA) used for effluent limit calculation.
}
