% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{cen_sum}
\alias{cen_sum}
\title{Sum up congener sample concentrations to create a composite parameter value.}
\usage{
cen_sum(qual, result, nd = c("<", "nd", "ND"))
}
\arguments{
\item{qual}{A character vector containing non-detect indicator strings, e.g., "<" or "ND". The strings used to indicate censored status can be edited in the "nd" argument.}

\item{result}{A numeric vector of concentration measurements.}

\item{nd}{A list indicating all the censoring flags included in the dataset. Defaults to "<", "nd", and "ND".}
}
\value{
dataframe with two columns, qual (character) and result (numeric)
}
\description{
Sum up congener sample concentrations to create a composite parameter value.
}
\examples{
## Example 1 - Some non-detects
result <- runif(5, 1, 10)
cen_result <- sample(c("<", ""), length(result), 
                     replace = TRUE, prob = c(0.2, 0.8))
cen_sum(cen_result, result)

## Example 2 - All non-detects
result <- runif(5, 1, 10)
cen_result <- sample(c("<", ""), length(result), 
                     replace = TRUE, prob = c(1, 0))
cen_sum(cen_result, result)

}
