% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{cv_adj}
\alias{cv_adj}
\title{Calculated adjusted coeficient of variation (CV).}
\usage{
cv_adj(qual, result, nd = c("<", "nd", "ND"), nd_adjustment = 0.5)
}
\arguments{
\item{qual}{A character vector containing non-detect indicator strings, e.g., "<" or "ND". The strings used to indicate censored status can be edited in the "nd" argument.}

\item{result}{A numeric vector of concentration measurements.}

\item{nd}{A list indicating all the censoring flags included in the dataset. Defaults to "<", "nd", and "ND".}

\item{nd_adjustment}{Adjustment factor for non-dectect values. Non-detect values (as indicated in qual vector) are multiplied by nd_adjustment factor; i.e. result * nd_adjustment. Typically, method detection limits or reporting limits are used as result values for non-detects.}
}
\value{
A numeric coefficient of variation (CV) value
}
\description{
Calculated adjusted coeficient of variation (CV).
}
