% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{project_mec}
\alias{project_mec}
\title{Calculate Projected Maximum Effluent Concentration (MEC)}
\usage{
project_mec(
  qual,
  result,
  nd = c("<", "nd", "ND"),
  percentile = 0.95,
  conf_level = 0.99,
  nd_adjustment = 0.5,
  simple_output = FALSE
)
}
\arguments{
\item{qual}{A character vector containing non-detect indicator strings, e.g., "<" or "ND". The strings used to indicate censored status can be edited in the "nd" argument.}

\item{result}{A numeric vector of concentration measurements.}

\item{nd}{A list indicating all the censoring flags included in the dataset. Defaults to "<", "nd", and "ND".}

\item{percentile}{Location on lognormal distribution to estimate the projected MEC.}

\item{conf_level}{Confidence level of projected estimate.}

\item{nd_adjustment}{Adjustment factor for non-dectect values. Non-detect values (as indicated in qual vector) are multiplied by nd_adjustment factor; i.e. result * nd_adjustment. Typically, method detection limits or reporting limits are used as result values for non-detects.}

\item{simple_output}{Logical value. If TRUE, the output columns are concatenated into a single character string (e.g., "<0.2") which can be useful for constructing summary tables.}
}
\value{
Dataframe with a qualifier column (character) and a MEC column (numeric).
}
\description{
Calculate Projected Maximum Effluent Concentration (MEC)
}
