% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{cen_sum}
\alias{cen_sum}
\title{Sum detected and non-detect concentrations.}
\usage{
cen_sum(qual, result, nd = c("<", "nd", "ND"))
}
\arguments{
\item{qual}{A character vector containing non-detect indicator strings, e.g., "<" or "ND". The strings used to indicate censored status can be edited in the "nd" argument.}

\item{result}{A numeric vector of concentration measurements.}

\item{nd}{A list indicating all the censoring flags included in the dataset. Defaults to "<", "nd", and "ND".}
}
\value{
dataframe with two columns, qual (character) and result (numeric)
}
\description{
Sum up congener sample concentrations to create a composite parameter value. Primarily intended as a helper function in fuse_samples.
}
